updated content

+++ routes/database.php
@@ -3,12 +3,16 @@
 function get_records($table,$cred) {
     $credentials = (object)$cred;
 
-    $conn = new mysqli_connect(
+    $conn = mysqli_connect(
         $credentials->host,
         $credentials->user,
         $credentials->password,
         $credentials->db
     );
+    if (!$conn) {
+        die("Connection failed: " . mysqli_connect_error());
+    }
+    
     $query = "SELECT * FROM $table";
     $result = mysqli_query($conn, $query);
     return mysqli_fetch_all($result, MYSQLI_ASSOC);
@@ -23,6 +27,9 @@ function get_record_by_email($table,$cred, $email) {
         $credentials->password,
         $credentials->db
     );
+    if (!$conn) {
+        die("Connection failed: " . mysqli_connect_error());
+    }
     $query = "SELECT * FROM $table WHERE email='$email'";
     $result = mysqli_query($conn, $query);
     return mysqli_fetch_all($result, MYSQLI_ASSOC);
@@ -37,6 +44,10 @@ function create_record($table,$cred, $data) {
         $credentials->password,
         $credentials->db
     );
+    if (!$conn) {
+        die("Connection failed: " . mysqli_connect_error());
+    }
+
     $fields = array_keys($data);
     $values = array_map(function($value) use ($conn) {
         return mysqli_real_escape_string($conn, $value);
@@ -54,6 +65,10 @@ function update_record($table,$cred, $email, $data) {
         $credentials->password,
         $credentials->db
     );
+    if (!$conn) {
+        die("Connection failed: " . mysqli_connect_error());
+    }
+
     $sets = [];
     foreach ($data as $field => $value) {
         $sets[] = "$field='".mysqli_real_escape_string($conn, $value)."'";
@@ -71,6 +86,9 @@ function delete_record($table,$cred, $id) {
         $credentials->password,
         $credentials->db
     );
+    if (!$conn) {
+        die("Connection failed: " . mysqli_connect_error());
+    }
     $query = "DELETE FROM $table WHERE id='$id'";
     return mysqli_query($conn, $query);
 }
